Create reference VM package
This adds a new reference package to the module which implements a
reference VM. This reference VM is completely unoptimized in favor of
being readable. This means that it will run really slowly, but allows
for simple validation of the code. Additionally, it makes it much
simpler to create a step function without worrying about performance
penalties.

A few other changes were made to implement this; namely, the cerr
package was renamed vmerr, and all the top-level errors were moved into
it so they could be shared between the two VM implementations. Further,
some functions in the top-level package for manipulating constant
arguments to opcodes were moved to the op package.

Finally, the ./cmd/runner code was adapted to use the reference VM, and
a new option --trace was added which prints the stack after each step.